backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-socketio>=5.12
httpx>=0.28
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.21
asyncpg>=0.29
python-dotenv>=1.0
openpyxl>=3.1
aiofiles>=24.1
playwright>=1.49
beautifulsoup4>=4.12
supabase>=2.0
